refactor: extract utils function
